Accept generateResource flag when updating roles.
generateResource is a flag that you can attach to roles
inside an identity when creating it, the flag basically
tells the api to restrict that role to only the identity's ID.
This extends this functionality to allow you to pass the flag
in on identity update as well.

Change generateResource to append, not overwrite.

Re-add identity.sysResourceRole assignment.

Check index of with === -1 instead of < 0.

Update error message.